use "DPoP proof JWT" or "DPoP proof" as terminology for the header JWT
one suggestion about terminology that I think we should consider before -02. I'd suggest we consider using a different term than "DPoP token" for the header JWT. And because language can be pretty sticky, a change sooner rather than later is probably better. There are tokens everywhere in OAuth and I found myself getting mixed up at times by the language when reading the doc.  I also suspect that, if this thing gets some adoption and usage, some people will use shorthand language to talk about things, like calling an access token that is bound to a DPoP public key a DPoP access token or just a DPoP token. Which is totally fine but lead to confusion about whether folks are referring to the aDPoP header JWT or the DPoP bound access token (i.e. "the dpop token didn't validate") in regular conversation about this stuff.

I rather like the term in the current -01 version of "DPoP proof JWT" or  just "DPoP proof". And think it works. It's the piece that's proving possession of the key (with respect to the http request). A server might bind a token to it or check a token that's bound to it or both. But this DPoP header is the proof part. So technically I think the language works and it helps distinguish this JWT from all the other tokens.