Including all optional parameters, and elaboration on examples.
Include `nonce` in examples.

Including decoded representations of DPoP proofs in more examples.

Included references to private key as well to emphasize redaction of
private key material. Many client implementers are not going to be as
comfortable with cryptography, and could use this being reiterated a few
more times (on the fence about this change honestly, but giving it a
shot).

Slightly more variation in nonce and jti parameters to show flexibility
in choosing values there.